requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
cachelib==0.1
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cryptography==3.3.1
cssselect==1.1.0
decorator==4.4.2
Deprecated==1.2.10
howdoi==2.0.7
idna==2.10
keep==2.10
lxml==4.6.2
pycparser==2.20
PyGithub==1.54
Pygments==2.7.3
PyJWT==1.7.1
pyquery==1.4.3
python-dotenv==0.15.0
python-telegram-bot==13.1
pytz==2020.4
requests==2.25.0
six==1.15.0
terminaltables==3.1.0
tornado==6.1
tzlocal==2.1
urllib3==1.26.2
wrapt==1.12.1
